Creating submodule and splitting main from package
