+++ cmd/new-api-example/go.mod
@@ -1,0 +1,3 @@
+module github.com/brunokim/causal-tree/new-api-sample
+
+go 1.18
